fix(server): Rename the hello entity to comply with the naming convention
