+++ server/src/entities/hello.entity.ts
